requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pymongo==4.5.0
python-dotenv==1.0.0